+++ src/scorer/scorer.ts
@@ -6,6 +6,7 @@ import Toot from '../api/objects/toot';
 import { DEFAULT_WEIGHTS } from "./weight_presets";
 import { SCORERS_CONFIG } from "../config";
 import { ScorerInfo, StringNumberDict, TootScore, WeightName } from "../types";
+import { sumValues } from "../helpers";
 
 const TIME_DECAY = WeightName.TIME_DECAY;
 
@@ -48,7 +49,6 @@ export default abstract class Scorer {
     // Add all the score into to a toot, including a final score
     static async decorateWithScoreInfo(toot: Toot, scorers: Scorer[]): Promise<void> {
         // console.debug(`decorateWithScoreInfo ${describeToot(toot)}: `, toot);
-        let rawScore = 1;
         const tootToScore = toot.reblog ?? toot;
         const rawScores = {} as StringNumberDict;
         const weightedScores = {} as StringNumberDict;
@@ -62,30 +62,28 @@ export default abstract class Scorer {
             const scoreValue = scores[i] || 0;
             rawScores[scorer.name] = scoreValue;
             weightedScores[scorer.name] = scoreValue * (userWeights[scorer.name] ?? 0);
-            rawScore += weightedScores[scorer.name];
+
+            if (tootToScore.isTrending()) {
+                weightedScores[scorer.name] *= (userWeights[WeightName.TRENDING] ?? 0);
+            }
         });
 
         // Multiple rawScore by time decay penalty to get a final value
         const timeDecay = userWeights[TIME_DECAY] || DEFAULT_WEIGHTS[TIME_DECAY];
-        const seconds = Math.floor((new Date().getTime() - new Date(tootToScore.createdAt).getTime()) / 1000);
-        const timeDecayMultiplier = Math.pow((1 + timeDecay), -1 * Math.pow((seconds / 3600), 2));
+        const timeDecayMultiplier = Math.pow((1 + timeDecay), -1 * Math.pow(tootToScore.ageInHours(), 2));
+        // Add 1 so that time decay multiplier works even with scorers giving 0s
+        const weightedScore = 1 + sumValues(weightedScores);
 
         tootToScore.scoreInfo = {
-            rawScore,
+            rawScore: 1 + sumValues(rawScores),  // Add 1 for same reason as weightedScore
             rawScores,
-            score: 0,
+            score: weightedScore * timeDecayMultiplier,
             timeDecayMultiplier,
             weightedScores,
+            weightedScore,
         } as TootScore;
 
-        // Trending toots usually have a lot of reblogs, likes, replies, etc. so they get disproportionately
-        // high scores. To adjust for this we hack a final adjustment to the score by multiplying by the
-        // trending weighting value.
-        if (tootToScore.isTrending()) {
-            tootToScore.scoreInfo.rawScore *= (userWeights[WeightName.TRENDING] ?? 0);
-        }
-
-        tootToScore.scoreInfo.score = tootToScore.scoreInfo.rawScore * timeDecayMultiplier;
-        toot.scoreInfo = tootToScore.scoreInfo;  // Copy the score info to the retoot if need be
+        // Copy the score info to the retoot if need be  // TODO: duping the score for retoots is a hack
+        toot.scoreInfo = tootToScore.scoreInfo;
     }
 };